+++ pages/about.page.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const About = () => {
+  return (
+    <div>
+      test
+    </div>
+  )
+}
+
+export default About
