+++ src/Layout/Homepage.jsx
@@ -2,12 +2,10 @@ import React from 'react';
 import 
     MobileNav
  from '../components/containers/Navbar';
-
  import DesktopNav from '../components/containers/Navbar/DesktopNav';
-
 import Hero from '../components/containers/Hero/index';
-
 import About from '../components/containers/About';
+
 
 const Homepage = () => {
   return (
