+++ src/components/containers/Hero/index.jsx
@@ -5,10 +5,10 @@ import {
     Image
 } from '@chakra-ui/react';
 import styles from './styles';
-import {hero} from '../../../constants/index';
+
 
 import AnvilLogo from '/AnvilLogo.png'
-// import Arrow from '../../Arrow';
+
 
 
 const index = () => {
@@ -17,8 +17,7 @@ const index = () => {
 <Text as='h1'>
 Crafting Beautiful <br/> Websites That <span>Inspire</span> <br/> & Convert
 </Text>
-<Text as='h2' >{hero[1].title}</Text>
-{/* <Text as='h2'>{hero[1].title}</Text> */}
+<Text as='h2' >Expert Web Development, Graphic Design & Coding Services</Text>
 
 <Box sx={styles.heroImage}>
   <Image src={AnvilLogo} alt='Site-Smiths-Anvil-Logo'
